figured out how to fix a big bug!! have to check and make sure a user is authorized to load pets, or any other model, and if they arent, you navigate back to the home page

+++ app/controllers/pet_controller.rb
@@ -1,19 +1,21 @@
 class PetController < ApplicationController
 
+    before_action :authorized
+
     def index
         user = Rescue.find_by(id: session[:user_id])
         if user
             pets = Pet.all
             render json: pets
         else
-            render json: { error: "Sign in to see the pets"}, status: :unauthorized
+            render json: { error: "ERROR: You must sign in to see the pets."}, status: :unauthorized
         end
     end
 
-    def show
-        pet = Pet.find(params[:id])
-        render json: pet
-    end
+    # def show
+    #     pet = Pet.find(params[:id])
+    #     render json: pet
+    # end
 
     def create
         user = Rescue.find_by(id: session[:user_id])
@@ -33,7 +35,7 @@ class PetController < ApplicationController
         user = Rescue.find_by(id: session[:user_id])
         if user
         pet = Pet.find_by(id: params[:id])
-        if pet.rescue_id == user.id
+        if pet && pet.rescue_id == user.id
             pet.destroy
             render json: { message: "Pet deleted successfully" }, status: :ok
         else
@@ -48,11 +50,11 @@ class PetController < ApplicationController
         user = Rescue.find_by(id: session[:user_id])
         if user 
             pet = Pet.find_by(id: params[:id])
-            if pet 
-                pet.update(pet_params)
+            pet.update(pet_params)
+            if pet && pet.valid?
                 render json: pet, status: :created
             else
-                render json: { error: ["Pet not found."]}, status: :not_found
+                render json: { error: "ERROR: Make sure all edits are filled out."}, status: :unprocessable_entity
             end
         else
             render json: {error: "ERROR: Not authorized. Must sign in to update, or update the pet your rescue is responsible for."}, status: :unauthorized
